Refactoring file names to temporary names
