+++ e2e/journeys/manageCases/caseProgression/completeOrderServiceOfApplication/statementOfService/statementOfService.ts
@@ -1,8 +1,8 @@
 import { Page } from "@playwright/test";
-import { Helpers } from "../../../../common/helpers";
-import { StatementOfService1Page } from "../../../../pages/manageCases/caseProgression/statementOfService/statementOfService1Page";
-import { StatementOfServiceSubmitPage } from "../../../../pages/manageCases/caseProgression/statementOfService/StatementOfServiceSubmitPage";
-import { WACaseWorkerActions } from "../../../../common/types.ts";
+import { Helpers } from "../../../../../common/helpers.ts";
+import { StatementOfService1Page } from "../../../../../pages/manageCases/caseProgression/statementOfService/statementOfService1Page.ts";
+import { StatementOfServiceSubmitPage } from "../../../../../pages/manageCases/caseProgression/statementOfService/StatementOfServiceSubmitPage.ts";
+import { WACaseWorkerActions } from "../../../../../common/types.ts";
 
 interface statementOfServiceParams {
   page: Page;
