+++ e2e/tests/manageCases/caseProgression/completeOrderServiceOfApplication/completeOrderServiceOfApplicationStatementOfService.spec.ts
@@ -3,7 +3,7 @@ import Config from "../../../../config";
 import createDaCitizenCourtNavCase from "../../../../common/createCaseHelper";
 import { Helpers } from "../../../../common/helpers";
 import config from "../../../../config";
-import { CompleteOrderServiceOfApplication } from "../../../../journeys/manageCases/caseProgression/completeOrderServiceOfApplication/completeOrderServiceOfApplication.ts";
+import { CompleteOrderServiceOfApplicationStatementOfService } from "../../../../journeys/manageCases/caseProgression/completeOrderServiceOfApplication/completeOrderServiceOfApplicationStatementOfService.ts";
 
 test.use({ storageState: Config.sessionStoragePath + "caseWorker.json" });
 
@@ -19,7 +19,7 @@ test.describe("Complete the Order task for DA Citizen case tests.", () => {
     page,
     browser,
   }): Promise<void> => {
-    await CompleteOrderServiceOfApplication.completeOrderServiceOfApplication({
+    await CompleteOrderServiceOfApplicationStatementOfService.completeOrderServiceOfApplication({
       page: page,
       accessibilityTest: false,
       yesNoSendToGateKeeper: true,
@@ -43,7 +43,7 @@ test.describe("Complete the Order task for DA Citizen case tests.", () => {
     page,
     browser,
   }): Promise<void> => {
-    await CompleteOrderServiceOfApplication.completeOrderServiceOfApplication({
+    await CompleteOrderServiceOfApplicationStatementOfService.completeOrderServiceOfApplication({
       page: page,
       accessibilityTest: false,
       yesNoSendToGateKeeper: true,
